Create module for Fox
