+++ src/fox/component.rs
@@ -1,0 +1,6 @@
+use bevy::{ecs::component::Component, math::Vec3};
+
+#[derive(Debug, Default, Component)]
+pub struct Fox {
+    target_position: Vec3,
+}
