+++ src/main.rs
@@ -1,9 +1,9 @@
 mod config;
+mod fox;
 mod voxel;
 
 use bevy::{
     app::{App, Startup, Update},
-    asset::AssetServer,
     core_pipeline::core_3d::Camera3dBundle,
     ecs::{
         change_detection::DetectChanges,
@@ -13,10 +13,9 @@ use bevy::{
         schedule::IntoSystemConfigs,
         system::{Commands, Local, Query, Res, ResMut},
     },
-    math::{Quat, Vec3},
+    math::Vec3,
     pbr::{AmbientLight, PointLight, PointLightBundle},
     render::color::Color,
-    scene::SceneBundle,
     time::Time,
     transform::components::Transform,
     utils::default,
@@ -25,6 +24,7 @@ use bevy::{
 #[cfg(feature = "with-inspector")]
 use bevy_inspector_egui::quick::ResourceInspectorPlugin;
 use bevy_rand::{plugin::EntropyPlugin, prelude::ChaCha8Rng};
+use fox::FoxPlugin;
 use serde::Deserialize;
 use voxel::SpawnVoxelEvent;
 
@@ -46,7 +46,8 @@ fn main() -> std::io::Result<()> {
     // Plugins
     app.add_plugins(DefaultPlugins)
         .add_plugins(EntropyPlugin::<ChaCha8Rng>::default())
-        .add_plugins(VoxelPlugin);
+        .add_plugins(VoxelPlugin)
+        .add_plugins(FoxPlugin);
 
     // Resources
     app.insert_resource(voxelcraft.world_configuration);
@@ -65,7 +66,6 @@ fn main() -> std::io::Result<()> {
 
     // Systems
     app.add_systems(Startup, create_camera)
-        .add_systems(Startup, spawn_fox)
         .add_systems(Update, spawn_voxels_on_timer)
         .run();
 
@@ -117,19 +117,6 @@ fn spawn_voxels_on_timer(
         *elapsed -= TIMER;
         event_writer.send(SpawnVoxelEvent);
     }
-}
-
-fn spawn_fox(mut commands: Commands, asset_server: Res<AssetServer>) {
-    let scene = asset_server.load("Fox.gltf#Scene0");
-    commands.spawn(SceneBundle {
-        scene,
-        transform: Transform {
-            translation: Vec3::from_array([0., 1., 0.]),
-            rotation: Quat::from_rotation_y(std::f32::consts::PI),
-            scale: Vec3::splat(1. / 32.),
-        },
-        ..default()
-    });
 }
 
 #[cfg(feature = "with-inspector")]
